refactor(kona-sp1): share super-range output builder
