+++ rust/kona/sp1/crates/ethereum/client/Cargo.toml
@@ -13,15 +13,24 @@ kona-sp1-client-utils.workspace = true
 kona-derive.workspace = true
 kona-driver.workspace = true
 kona-executor.workspace = true
-kona-genesis.workspace = true
-kona-interop.workspace = true
+kona-genesis = { workspace = true, features = ["serde"] }
+kona-interop = { workspace = true, features = ["serde"] }
 kona-preimage.workspace = true
 kona-proof.workspace = true
+kona-proof-interop.workspace = true
+kona-registry.workspace = true
 
 # general
 anyhow.workspace = true
 async-trait.workspace = true
+serde_json.workspace = true
 spin.workspace = true
+
+# alloy
+alloy-chains.workspace = true
+alloy-consensus.workspace = true
+alloy-primitives.workspace = true
+alloy-rlp.workspace = true
 
 [lints]
 workspace = true
