+++ rust/kona/sp1/crates/ethereum/client/src/lib.rs
@@ -1,3 +1,4 @@
 //! Ethereum sp1 proof client implementation.
 
 pub mod executor;
+pub mod super_range;
